obj_gen: generate pure random value buffer
For some reason, memtier didn't generate a pure random value buffer when
random data is needed.
This is not good for us for several reasons (assuming we are running with
compression):
1. The effective capacity is bigger than expected.
2. We might not reach the desired WA => can't run performance tests.
3. We would like to have a better control how much the generated data is
compressed.

So, we change the implementation, so the whole value buffer is read from
the /dev/urandom, hence getting a pure random data.

Issue: LBM1-3458