Binary memcache protocol: added multi get support to request parsing
Protocol response now holds a list of return values.
If a response is a part of a GETK multi get sequence ,
we will accumulate all the returning values and keys in the response values list.
only when the sequence finisher command (GETKQ/NOOP) will arrive ,
the response parsing will be finished and handled at the client level.

ISSUE: LBM1-3649