Protocol: response values list now holds key and key length for each value
Added key and key length to the list node of the values list in the protocol resonse class.
this ws done so we will be able to verify multi key get requests,
as the values are not returning in order we need to save the key returned for each value.

ISSUE: LMB1-3876