Protocol/memcache_binary: add key not faound response status to errors list.
Added the no key found status to the errors list in the protocol parsing
method.
If we start a get only workload on memtier without first setting the
keys we will get responses from the server containing the key not found
error status.
This status was not marked as an error in the protocol parsing method,
so we would have goten a segfault error when trying to manipulate empty
buffers.

ISSUE: LBM1-4057