Client: handle timeout race when updating event after writing to socket
after writing from the write buffer to socket or reading from socket we
update the event with read/write commands according to the test status.

if the write buffer still have more data in it we will scheduale write event,
right after checking the remaining write buffer length we used toassert if the test is finished.
this is wrong.

memtier tests can either be limited by timeout or by a requests number threshold not both,
in the case of requests number threshold if the test is finished and the write buffer
isnt empty an error occured and we should abort the test.
in the case of the test finished due to timeout reached even if the write buffer
is not empty we should just exit gracefully so stats wont be affected.

Issue: LBM1-4819